zynqmp: Define default DDR location if DEBUG enabled
This patch defines a default DDR location to which ATF
has to compiled if DEBUG option is enabled. This is required
now, as the ATF cant fit in OCM with DEBUG option enabled.
The deafult value is 0x1000 and can be used till 0x7ffff.
User can still override as per wish/requirement using
current commandline options.

Signed-off-by: Siva Durga Prasad Paladugu <sivadur@xilinx.com>